need to do backward compatibility.

+++ image_upload/views.py
@@ -28,24 +28,29 @@ from django.shortcuts import get_object_or_404
 from django import forms, http
 from django.shortcuts import render
 from django.contrib.auth.decorators import login_required
+from django.core.files.base import ContentFile
 from django.views.generic import (
         CreateView, DeleteView, ListView)
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.utils.translation import ugettext_lazy as _, string_concat, ugettext
 from django.db import transaction
 
-from course.models import Course, FlowPageData, Participation
+from course.models import Course, FlowPageData, Participation, FlowSession
 from course.utils import course_view
 
 from image_upload.serialize import serialize
 from image_upload.utils import get_page_image_behavior, ImageOperationMixin
 from image_upload.models import FlowPageImage
+from image_upload.storages import UserImageStorage
 
 from jsonview.decorators import json_view
 from jsonview.exceptions import BadRequest
 from braces.views import JSONResponseMixin
 import json
 from PIL import Image
+from io import BytesIO
+
+storage = UserImageStorage()
 
 
 def is_course_staff(pctx):
@@ -98,9 +103,6 @@ class ImageCreateView(LoginRequiredMixin, ImageOperationMixin, JSONResponseMixin
         self.object = form.save(commit=False)
         self.object.creator = self.request.user
 
-        from django.core.files.uploadedfile import TemporaryUploadedFile
-        print(type(self.object.file))
-        print(isinstance(self.object.file, TemporaryUploadedFile))
         self.object.save()
 
         course_identifier = self.kwargs["course_identifier"]
@@ -147,8 +149,11 @@ class ImageDeleteView(LoginRequiredMixin, ImageOperationMixin, DeleteView):
             from django.core.exceptions import PermissionDenied
             raise PermissionDenied(_("may not delete other people's image"))
         else:
-            self.object.delete()
-            response = http.JsonResponse(True, safe=False)
+            if storage.is_temp_image(self.object.file.path):
+                self.object.delete()
+                response = http.JsonResponse(True, safe=False)
+            else:
+                response = http.JsonResponse(False, safe=False)
             response['Content-Disposition'] = 'inline; filename=files.json'
             response['Content-Type'] = 'text/plain'
             return response
@@ -161,21 +166,79 @@ class ImageListView(LoginRequiredMixin, JSONResponseMixin, ListView):
 
     def get_queryset(self):
         flow_session_id = self.kwargs["flow_session_id"]
+        flow_session = get_object_or_404(FlowSession, id=int(flow_session_id))
         ordinal = self.kwargs["ordinal"]
+        course_identifier = self.kwargs["course_identifier"]
+        from course.utils import CoursePageContext
+        pctx = CoursePageContext(self.request, course_identifier)
+        prev_visit_id = self.request.GET.get("visit_id")
 
-        try:
-            fpd = FlowPageData.objects.get(
-                    flow_session=flow_session_id, ordinal=ordinal)
-        except ValueError:
 
-            # in sandbox
-            if flow_session_id == "None" or ordinal == "None":
-                return None
+        answer_visit = None
+        answer_data = None
+        if prev_visit_id is not None:
+            try:
+                prev_visit_id = int(prev_visit_id)
+            except ValueError:
+                from django.core.exceptions import SuspiciousOperation
+                raise SuspiciousOperation("non-integer passed for 'visit_id'")
+        from course.utils import FlowPageContext
+        from course.flow import get_prev_answer_visits_qset
+        fpctx = FlowPageContext(
+            repo=pctx.repo,
+            course=pctx.course,
+            flow_id=flow_session.flow_id,
+            ordinal=int(ordinal),
+            flow_session=flow_session,
+            participation=flow_session.participation,
+            request=self.request
+        )
 
-        return FlowPageImage.objects\
-                .filter(flow_session=flow_session_id)\
-                .filter(image_page_id=fpd.page_id)\
-                .order_by("order","pk")
+        prev_answer_visits = list(
+                get_prev_answer_visits_qset(fpctx.page_data))
+
+
+        # {{{ fish out previous answer_visit
+
+        viewing_prior_version = False
+        if prev_answer_visits and prev_visit_id is not None:
+            answer_visit = prev_answer_visits[0]
+
+            for ivisit, pvisit in enumerate(prev_answer_visits):
+                if pvisit.id == prev_visit_id:
+                    answer_visit = pvisit
+                    if ivisit > 0:
+                        viewing_prior_version = True
+
+                    break
+
+            prev_visit_id = answer_visit.id
+
+        elif prev_answer_visits:
+            answer_visit = prev_answer_visits[0]
+            prev_visit_id = answer_visit.id
+
+        else:
+            answer_visit = None
+
+        # }}}
+
+        if not answer_visit:
+            return None
+
+        answer_data = answer_visit.answer
+        pk_list = answer_data.get("answer", None)
+        if not pk_list:
+            return None
+
+        # Creating a QuerySet from a list while preserving order using Django
+        # https://codybonney.com/creating-a-queryset-from-a-list-while-preserving-order-using-django/
+        clauses = ' '.join(['WHEN id=%s THEN %s' % (pk, i) for i, pk in enumerate(pk_list)])
+        ordering = 'CASE %s END' % clauses
+        queryset = FlowPageImage.objects.filter(pk__in=pk_list).extra(
+            select={'ordering': ordering}, order_by=('ordering',))
+
+        return queryset
 
     def render_to_response(self, context, **response_kwargs):
         queryset = self.get_queryset()
@@ -283,6 +346,7 @@ def image_crop_modal(pctx, flow_session_id, ordinal, pk):
              'owner': file.creator
             })
 
+
 @json_view
 @login_required
 @transaction.atomic
@@ -311,7 +375,6 @@ def image_crop(pctx, flow_session_id, ordinal, pk):
             string_concat(_('File not found.'),
                           _('Please upload the image first.')))
 
-    image_modified_path = crop_instance.get_random_filename()
 
     if not request.POST:
         return {}
@@ -329,46 +392,62 @@ def image_crop(pctx, flow_session_id, ordinal, pk):
         raise CropImageError(_('There are errors, please refresh the page or try again later'))
 
     try:
-        image_orig = Image.open(image_orig_path)
+        new_image = Image.open(image_orig_path)
     except IOError:
         raise CropImageError(_('There are errors，please re-upload the image'))
+    image_format = new_image.format
 
     if rotate != 0:
         # or it will raise "AttributeError: 'NoneType' object has no attribute 'mode' error
         # in pillow 3.3.0
-        image_orig = image_orig.rotate(-rotate, expand=True)
+        new_image = new_image.rotate(-rotate, expand=True)
 
     box = (x, y, x+width, y+height)
-    image_orig = image_orig.crop(box)
+    new_image = new_image.crop(box)
 
-    try:
-        if image_orig.mode != "RGB":
-            image_orig = image_orig.convert("RGB")
-        image_orig.save(image_modified_path)
-    except (OSError, IOError):
-        raise CropImageError(_('There are errors, please refresh the page or try again later'))
+    if new_image.mode != "RGB":
+        new_image = new_image.convert("RGB")
 
     from relate.utils import as_local_time, local_now
     import datetime
 
+    new_image_file_last_modified = crop_instance.creation_time
     if not course_staff_status:
-        if local_now() < as_local_time(
-                crop_instance.creation_time + datetime.timedelta(minutes=5)):
-            crop_instance.file_last_modified = crop_instance.creation_time = local_now()
+        if local_now() > as_local_time(
+                        crop_instance.creation_time + datetime.timedelta(minutes=5)):
+            new_image_file_last_modified = local_now()
 
-        else:
-            crop_instance.file_last_modified = local_now()
+    new_image_io = BytesIO()
+    new_image.save(new_image_io, format=image_format)
 
-    crop_instance.file = image_modified_path
-    crop_instance.save()
+    new_instance = FlowPageImage()
+    new_instance.creator=crop_instance.creator
+    new_instance.file.save(
+        name=crop_instance.slug,
+        content=ContentFile(new_image_io.getvalue()),
+        save=False
+    )
+
+    new_instance.slug=crop_instance.slug
+    new_instance.creation_time=crop_instance.creation_time
+    new_instance.file_last_modified=new_image_file_last_modified
+    new_instance.course=crop_instance.course
+    new_instance.flow_session=crop_instance.flow_session
+    new_instance.image_page_id=crop_instance.image_page_id
+    new_instance.is_image_textify=crop_instance.is_image_textify
+    new_instance.image_text=crop_instance.image_text
+    new_instance.image_data=crop_instance.image_data
+    new_instance.use_image_data=crop_instance.use_image_data
 
     try:
-        import os
-        os.remove(image_orig_path)
-    except:
-        pass
-
-    new_instance = FlowPageImage.objects.get(id=pk)
+        new_instance.save()
+    except (OSError, IOError) as e:
+        raise CropImageError(string_concat(
+            _('There are errors, please refresh the page or try again later'),
+            "--%s:%s." % (type(e).__name__, str(e))
+        ))
+    finally:
+        new_image.close()
 
     response_file = serialize(request, new_instance, 'file')
